Unify modules into single script with Korean comments
