--- gui/__init__.py
@@ -1,3 +1,0 @@
-from .main import Application
-
-__all__ = ["Application"]
