--- model_training/__init__.py
@@ -1,3 +1,0 @@
-from .trainer import Trainer
-
-__all__ = ["Trainer"]
