--- news_fetcher/__init__.py
@@ -1,3 +1,0 @@
-from .fetcher import get_recent_news
-
-__all__ = ["get_recent_news"]
