--- video_generator/__init__.py
@@ -1,3 +1,0 @@
-from .generator import VideoGenerator
-
-__all__ = ["VideoGenerator"]
