--- youtube_downloader/__init__.py
@@ -1,3 +1,0 @@
-from .downloader import download_korean_shorts
-
-__all__ = ["download_korean_shorts"]
