--- youtube_uploader/__init__.py
@@ -1,3 +1,0 @@
-from .uploader import upload_video
-
-__all__ = ["upload_video"]
